Fix MockPositionInjectionTest: drain stale PUB messages before position assertion, fix FLOAT32 via.f32 bug
Agent-Logs-Url: https://github.com/raeedcho/hapticore/sessions/d3ebfd19-e979-4d8a-9d75-2329e28de270

+++ cpp/haptic_server/src/main.cpp
@@ -83,7 +83,7 @@ std::optional<Vec3> parse_vec3_param(const msgpack::object& params, const char* 
                 if (elem.type == msgpack::type::FLOAT64) {
                     v[j] = elem.via.f64;
                 } else if (elem.type == msgpack::type::FLOAT32) {
-                    v[j] = static_cast<double>(elem.via.f64);
+                    v[j] = static_cast<double>(elem.via.f32);
                 } else if (elem.type == msgpack::type::POSITIVE_INTEGER) {
                     v[j] = static_cast<double>(elem.via.u64);
                 } else if (elem.type == msgpack::type::NEGATIVE_INTEGER) {
